Updated "report" [has the current submitted version for ACML2013]
Committed the downloaded "pdb_files" folder and the
extracted "extracted_pdb_files" folder

+++ superposition/extract.py
@@ -2,14 +2,14 @@ import sys, codecs, os, re
 if sys.stdout.encoding is None:
         sys.stdout = codecs.open("/dev/stdout", "w", 'utf-8')
         
-fr = open('test/lcb_results.txt','r')        
-file_download = open('download.sh','w')
+fr = open('test/lcb_results_2.txt','r')        
+file_download = open('download_2.sh','w')
 pdb_ids = []
-file_extract = open('extract.sh','w')
+file_extract = open('extract_2.sh','w')
 line = fr.readline()
 line_count = 0
 pdb_files_count = 0
-file_simulate = open('simulate.sh','w')
+file_simulate = open('simulate_2.sh','w')
 while line != '':
   x = line.strip('\n')
   y = x.split()
@@ -20,14 +20,15 @@ while line != '':
   pdb_file_name = pdb_id + '.pdb.gz'
   pdb_file_path = 'test/pdb_files/' + pdb_file_name
   if pdb_id not in pdb_ids:
-    pdb_ids.append(pdb_id)
-    download_cmd = 'wget http://www.rcsb.org/pdb/files/' + pdb_file_name 
-    file_download.write(download_cmd+'\n')
-    mv_cmd = 'mv ' + pdb_file_name + ' ' + pdb_file_path
-    file_download.write(mv_cmd+'\n')
-    rm_cmd = 'rm ' + pdb_file_name
-    file_download.write(rm_cmd+'\n');
-    pdb_files_count += 1
+    if not os.path.isfile(pdb_file_path):
+      pdb_ids.append(pdb_id)
+      download_cmd = 'wget http://www.rcsb.org/pdb/files/' + pdb_file_name 
+      file_download.write(download_cmd+'\n')
+      mv_cmd = 'mv ' + pdb_file_name + ' ' + pdb_file_path
+      file_download.write(mv_cmd+'\n')
+      rm_cmd = 'rm ' + pdb_file_name
+      file_download.write(rm_cmd+'\n');
+      pdb_files_count += 1
 
   # construct the command to extract from the pdb file
   chain = y[1]
@@ -41,8 +42,8 @@ while line != '':
   extract_cmd += ' > ' + extracted_file_path
   file_extract.write(extract_cmd+'\n')
   line_count += 1
-  simulate_cmd = './superimpose --proteins test/2IC7_test.pdb ' + extracted_file_path + ' '
-  simulate_cmd += '--accept 0.001 --translate 0.001 --rotate 0.05'
+  simulate_cmd = './superimpose --proteins test/query.pdb ' + extracted_file_path + ' '
+  simulate_cmd += '--accept 0.001 --translate 0.001 --rotate 0.05 --estimate general'
   file_simulate.write(simulate_cmd+'\n');
   print line_count
   line = fr.readline()
